Add files via upload
HTML and JS based local pictures viewer, by Dr. ChenGuang
Click the button and select one or multiple pictures with Shift-key or Ctrl-key and mouse-click. (Ctrl-a to select all the pictures in the folder)
Double click on the image to show in full-screen, Esc-key to exit full-screen.
Drag with left-mouse-button to move the picture.
Mouse-wheel to show the next picture.
Mouse-wheel with left-mouse-button held to rotate the picture.
Mouse-wheel with right-mouse-button held to zoom the picture.
Click mouse-wheel to move the center of the image to the point clicked.